+++ packages/edit/i18n/en.js
@@ -6,6 +6,7 @@ i18n.map('en', {
 	btn_edit: "Edit",
 	btn_del: "Delete",
 	btn_ren: "Rename",
+	btn_reload: "Reload",
 
 	label_editplaces: "Created places",